feat: apply feedbacks (#3349)
* refactor: changing the template structure


* chore: hide overflow-x scroll


* chore: update translation


* feat: apply selected favorite items


---------
