+++ apps/web/src/common/modules/navigations/gnb/GNBNavigationRail.vue
@@ -224,7 +224,7 @@ const refinedMenuList = (list, value) => {
 
 <style scoped lang="postcss">
 .g-n-b-navigation-rail {
-    @apply flex-col items-start bg-white border-r overflow-y-auto;
+    @apply flex-col items-start bg-white border-r overflow-y-auto overflow-x-hidden;
     top: $gnb-toolbox-height;
     width: $gnb-navigation-rail-max-width;
     height: calc(100% - $gnb-toolbox-height);
