+++ packages/language-pack/en.json
@@ -3577,7 +3577,7 @@
 			"ACTION_DELETE_THIS_GROUP": "Delete selected group",
 			"ACTION_EDIT_GROUP_NAME": "Edit group name",
 			"ADD_SERVICE_ACCOUNT": "Add service account",
-			"ALL_PROJECTS": "Project",
+			"ALL_PROJECTS": "All projects",
 			"ALT_E_CREATE_PROJECT": "Failed to Create Project",
 			"ALT_E_CREATE_PROJECT_GROUP": "Failed to Create Project Group",
 			"ALT_E_DELETE_PROJECT_GROUP": "Failed to Delete Project Group",
