+++ packages/language-pack/ja.json
@@ -3577,7 +3577,7 @@
 			"ACTION_DELETE_THIS_GROUP": "選択したグループ削除",
 			"ACTION_EDIT_GROUP_NAME": "選択したグループ編集",
 			"ADD_SERVICE_ACCOUNT": "サービスアカウント追加",
-			"ALL_PROJECTS": "プロジェクト",
+			"ALL_PROJECTS": "すべてのプロジェクト",
 			"ALT_E_CREATE_PROJECT": "プロジェクト作成に失敗",
 			"ALT_E_CREATE_PROJECT_GROUP": "プロジェクトグループ作成失敗",
 			"ALT_E_DELETE_PROJECT_GROUP": "プロジェクトグループの削除に失敗",
